feat: update changes

+++ schemas/concepto.py
@@ -6,16 +6,16 @@ class Concepto(BaseModel):
     clave: str = "0123456789"
     nombre: str = "Concepto"
     unidad: str = "m2"
-    precio: int = 100
+    precio: float = 100
     cantidad: int = 200
-    importe: int = precio * cantidad
+    importe: float = precio * cantidad
     descripcion: Optional[str] = None
 
 
 class CreaConcepto(BaseModel):
     nombre: str = "Concepto"
     unidad: str = "m2"
-    precio: int = 100
+    precio: float = 100
     cantidad: int = 200
     descripcion: Optional[str]
 
@@ -23,6 +23,6 @@ class CreaConcepto(BaseModel):
 class ModificaConcepto(BaseModel):
     nombre: Optional[str] = "Modificado"
     unidad: Optional[str] = "m3"
-    precio: Optional[int] = 200
+    precio: Optional[float] = 200
     cantidad: Optional[int] = 300
     descripcion: Optional[str] = "Agrega descripcion"
